MoAr better naming.  Move workspace snapshot and test to model from animations.
