+++ animations/src/test/java/com/chiralbehaviors/CoRE/meta/workspace/WorkspaceSnapshotTest.java
@@ -24,28 +24,22 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.util.Arrays;
-import java.util.List;
 
 import javax.persistence.EntityManager;
 
 import org.junit.Test;
 
-import com.chiralbehaviors.CoRE.Ruleform;
-import com.chiralbehaviors.CoRE.agency.Agency;
 import com.chiralbehaviors.CoRE.json.CoREModule;
 import com.chiralbehaviors.CoRE.meta.Model;
 import com.chiralbehaviors.CoRE.meta.models.AbstractModelTest;
 import com.chiralbehaviors.CoRE.meta.models.ModelImpl;
 import com.chiralbehaviors.CoRE.meta.workspace.dsl.WorkspaceImporter;
 import com.chiralbehaviors.CoRE.product.Product;
-import com.chiralbehaviors.CoRE.workspace.WorkspaceAuthorization;
 import com.chiralbehaviors.CoRE.workspace.WorkspaceSnapshot;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -159,87 +153,5 @@ public class WorkspaceSnapshotTest extends AbstractModelTest {
             myEm.close();
         }
 
-    }
-
-    @Test
-    public void testSerializeWorkspaceSnapshot() throws Exception {
-        Agency pseudoScientist = new Agency("Behold the Pseudo Scientist!");
-        pseudoScientist.setUpdatedBy(pseudoScientist);
-        Product definingProduct = new Product("zee product", pseudoScientist);
-        WorkspaceAuthorization auth = new WorkspaceAuthorization(definingProduct,
-                                                                 definingProduct,
-                                                                 pseudoScientist,
-                                                                 em);
-        auth = new WorkspaceAuthorization(pseudoScientist, definingProduct,
-                                          pseudoScientist, em);
-        WorkspaceSnapshot snapshot = new WorkspaceSnapshot(definingProduct,
-                                                           Arrays.asList(auth),
-                                                           model.getEntityManager());
-        snapshot.retarget(em);
-        em.flush();
-        WorkspaceSnapshot retrieved = new WorkspaceSnapshot(definingProduct,
-                                                            em);
-        ByteArrayOutputStream os = new ByteArrayOutputStream();
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.registerModule(new CoREModule());
-        mapper.writeValue(os, retrieved);
-        ByteArrayInputStream is = new ByteArrayInputStream(os.toByteArray());
-        WorkspaceSnapshot deserialized = mapper.readValue(is,
-                                                          WorkspaceSnapshot.class);
-        assertEquals(2, deserialized.getRuleforms()
-                                    .size());
-        assertEquals(pseudoScientist, deserialized.getRuleforms()
-                                                  .get(0));
-
-    }
-
-    @Test
-    public void testWorkspaceSnapshot() {
-        Agency pseudoScientist = new Agency("Behold the Pseudo Scientist!");
-        pseudoScientist.setUpdatedBy(pseudoScientist);
-        em.persist(pseudoScientist);
-        Product definingProduct = new Product("zee product", pseudoScientist);
-        em.persist(definingProduct);
-        new WorkspaceAuthorization(pseudoScientist, definingProduct,
-                                   pseudoScientist, em);
-        new WorkspaceAuthorization(definingProduct, definingProduct,
-                                   pseudoScientist, em);
-        Product aLeak = new Product("Snowden", kernel.getCore());
-        em.persist(aLeak);
-        new WorkspaceAuthorization(aLeak, definingProduct, pseudoScientist, em);
-        em.getTransaction()
-          .commit();
-        em.getTransaction()
-          .begin();
-        WorkspaceSnapshot snapshot = new WorkspaceSnapshot(definingProduct,
-                                                           model.getEntityManager());
-        em.getTransaction()
-          .rollback();
-        em.getTransaction()
-          .begin();
-        assertEquals("Did not find the leak!", 1, snapshot.getFrontier()
-                                                          .size());
-        Ruleform mole = snapshot.getFrontier()
-                                .get(0);
-        assertEquals("Imposter!", kernel.getCore(), mole);
-        em.getTransaction()
-          .rollback();
-        em.getTransaction()
-          .begin();
-        List<WorkspaceAuthorization> retrieved = WorkspaceSnapshot.getAuthorizations(definingProduct,
-                                                                                     em);
-        assertEquals(3, retrieved.size());
-        for (WorkspaceAuthorization a : retrieved) {
-            Ruleform ruleform = a.getRuleform(em);
-            if (ruleform instanceof Agency) {
-                assertEquals(pseudoScientist, ruleform);
-            } else if (!ruleform.equals(definingProduct)) {
-                assertEquals(aLeak, ruleform);
-                assertEquals("compromised!", kernel.getCore()
-                                                   .getName(),
-                             ruleform.getUpdatedBy()
-                                     .getName());
-            }
-        }
     }
 }
