+++ create-database/pom.xml
@@ -87,4 +87,6 @@
             </build>
         </profile>
     </profiles>
+    <name>Create Database</name>
+    <description>Perform the update on the DB</description>
 </project>
